refactor: remove notion_df.util
