+++ notion_df/contents.py
@@ -12,7 +12,7 @@ from notion_df.object.constant import BlockColor, CodeLanguage
 from notion_df.object.file import File
 from notion_df.object.misc import Icon
 from notion_df.object.rich_text import RichText, Span
-from notion_df.util.collection import FinalDict
+from notion_df.core.collection import FinalDict
 
 if TYPE_CHECKING:
     from notion_df.object.data import BlockData
