+++ notion_df/request/block.py
@@ -7,7 +7,7 @@ from uuid import UUID
 from notion_df.contents import BlockContents, serialize_block_contents_list
 from notion_df.core.request import SingleRequestBuilder, RequestSettings, Version, Method, PaginatedRequestBuilder
 from notion_df.object.data import BlockData
-from notion_df.util.collection import DictFilter
+from notion_df.core.collection import DictFilter
 
 
 @dataclass
