+++ notion_df/request/page.py
@@ -11,7 +11,7 @@ from notion_df.object.data import PageData
 from notion_df.object.file import ExternalFile
 from notion_df.object.misc import Icon, PartialParent
 from notion_df.property import PageProperties, Property, property_registry, PagePropertyValueT
-from notion_df.util.collection import DictFilter
+from notion_df.core.collection import DictFilter
 
 
 @dataclass
