+++ workflow/core/action.py
@@ -8,7 +8,7 @@ from typing import Iterable, Any, final, Callable, TypeVar, Optional
 from loguru import logger
 
 from notion_df.entity import Page, search_by_title
-from notion_df.util.misc import repr_object
+from notion_df.core.definition import repr_object
 from notion_df.variable import print_width
 from workflow.block_enum import is_template
 from workflow.core.entrypoint import entrypoint, WorkflowRecord, WorkflowSkipException
